Fixed a bunch of memory shit

+++ arch/i386/include/stdio.h
@@ -1,6 +1,6 @@
 #ifndef __STDIO_H
 #define __STDIO_H 
 void stdout_write(void *buf,int offset,int n);
-void putx(int n);
+void putx(unsigned long n);
 void blink();
 #endif
